tool_executor/requirements.txt
# Containerized Tool Executor Dependencies

# Gemini AI SDK
google-genai>=1.0.0

# Docker integration
docker>=7.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Console output
colorama>=0.4.6
